Social Web Mining - Twitter
The code gets the Twitter trends for 5 locations of interest today, by
using the countries WOEID. I used a modified version of pygmaps,
available under Issue 5 at https://code.google.com/p/pygmaps/. This
modified version enables the use of labels for the maps pinpoints. When
the user hovers over the pinpoint I display the top 5 trending topics on
Twitter for the area indicated by the pin point on the map.